Refactor Chrome Proxy Service
