+++ dwds/lib/src/chrome_proxy_service.dart
@@ -11,9 +11,21 @@ import 'package:pub_semver/pub_semver.dart' as semver;
 import 'package:vm_service_lib/vm_service_lib.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
-import 'dart_uri.dart';
 import 'debugger.dart';
 import 'helpers.dart';
+import 'inspector.dart';
+
+/// A handler for application assets, e.g. Dart sources.
+typedef AssetHandler = Future<String> Function(String);
+
+/// Adds [event] to the stream with [streamId] if there is anybody listening
+/// on that stream.
+typedef StreamNotify = void Function(String streamId, Event event);
+
+/// Returns the [Inspector] for the current tab.
+///
+/// This may be null during a hot restart or page refresh.
+typedef CurrentInspector = Inspector Function();
 
 /// A proxy from the chrome debug protocol to the dart vm service protocol.
 class ChromeProxyService implements VmServiceInterface {
@@ -30,40 +42,36 @@ class ChromeProxyService implements VmServiceInterface {
   final ChromeTab _tab;
 
   /// The connection with the chrome debug service for the tab.
+  // TODO(grouma) - This should be class private.
   final WipConnection tabConnection;
 
-  /// A handler for application assets, e.g. Dart sources.
-  final Future<String> Function(String) assetHandler;
-
-  /// The isolate for the current tab.
-  ///
-  /// This may be null during a hot restart or page refresh.
-  Isolate _isolate;
-
-  Isolate get isolate => _isolate;
+  final AssetHandler _assetHandler;
 
   /// Provides debugger-related functionality.
-  Debugger debugger;
+  Debugger _debugger;
 
-  /// Fields that are specific to the current [_isolate].
-  ///
-  /// These need to get cleared whenever [destroyIsolate] is called.
-  // TODO(grouma) - Subclass Isolate to group these together. This will ensure
-  // that they are properly cleaned up and documented.
-  final _classes = <String, Class>{};
-  final _scriptRefs = <String, ScriptRef>{};
-  final _serverPathToScriptRef = <String, ScriptRef>{};
-  final _libraries = <String, Library>{};
-  final _libraryRefs = <String, LibraryRef>{};
+  Inspector _inspector;
+  Inspector _currentInspector() => _inspector;
+
   StreamSubscription<ConsoleAPIEvent> _consoleSubscription;
 
   ChromeProxyService._(
-      this._vm, this._tab, this.tabConnection, this.assetHandler);
+    this._vm,
+    this._tab,
+    this.tabConnection,
+    this._assetHandler,
+  ) {
+    _debugger = Debugger(
+      _assetHandler,
+      tabConnection,
+      _streamNotify,
+      _currentInspector,
+      uri,
+    );
+  }
 
-  static Future<ChromeProxyService> create(
-      ChromeConnection chromeConnection,
-      Future<String> Function(String) assetHandler,
-      String appInstanceId) async {
+  static Future<ChromeProxyService> create(ChromeConnection chromeConnection,
+      AssetHandler assetHandler, String appInstanceId) async {
     ChromeTab appTab;
     for (var tab in await chromeConnection.getTabs()) {
       if (tab.url.startsWith('chrome-extensions:')) continue;
@@ -96,67 +104,46 @@ class ChromeProxyService implements VmServiceInterface {
   }
 
   Future<Null> _initialize() async {
-    debugger = Debugger(this);
-    await debugger.initialize();
+    await _debugger.initialize();
   }
 
   /// The root URI at which we're serving.
   String get uri => _tab.url;
 
-  /// Creates a new [_isolate].
+  /// Creates a new isolate.
   ///
   /// Only one isolate at a time is supported, but they should be cleaned up
   /// with [destroyIsolate] and recreated with this method there is a hot
   /// restart or full page refresh.
   Future<void> createIsolate() async {
-    if (_isolate != null) {
+    if (_inspector?.isolate != null) {
       throw UnsupportedError(
           'Cannot create multiple isolates for the same app');
     }
 
-    var id = createId();
-    var isolate = Isolate()
-      ..id = id
-      ..number = id
-      ..name = '$uri:main()'
-      ..runnable = true
-      ..breakpoints = []
-      ..libraries = []
-      ..extensionRPCs = [];
-    var isolateRef = toIsolateRef(isolate);
-    isolate.pauseEvent = Event()
-      ..kind = EventKind.kResume
-      ..isolate = isolateRef;
+    _inspector = Inspector(
+      createId(),
+      tabConnection,
+      _assetHandler,
+      _debugger,
+      uri,
+    );
 
-    for (var library in await _getLibraryNames()) {
-      isolate.libraries.add(LibraryRef()
-        ..id = library
-        ..name = library
-        ..uri = library);
-    }
+    await _inspector.initialize();
 
-    // TODO: Something more robust here, right now we rely on the 2nd to last
-    // library being the root one (the last library is the bootstrap lib).
-    isolate.rootLib = isolate.libraries[isolate.libraries.length - 2];
-
-    isolate.extensionRPCs.addAll(await _getExtensionRpcs());
-
-    for (var libraryRef in isolate.libraries) {
-      _libraryRefs[libraryRef.id] = libraryRef;
-    }
+    var isolateRef = toIsolateRef(_inspector.isolate);
 
     // Listen for `registerExtension` and `postEvent` calls.
     _setUpChromeConsoleListeners(isolateRef);
 
     _vm.isolates.add(isolateRef);
-    _isolate = isolate;
 
-    streamNotify(
+    _streamNotify(
         'Isolate',
         Event()
           ..kind = EventKind.kIsolateStart
           ..isolate = isolateRef);
-    streamNotify(
+    _streamNotify(
         'Isolate',
         Event()
           ..kind = EventKind.kIsolateRunnable
@@ -165,8 +152,8 @@ class ChromeProxyService implements VmServiceInterface {
     // TODO: We shouldn't need to fire these events since they exist on the
     // isolate, but devtools doesn't recognize extensions after a page refresh
     // otherwise.
-    for (var extensionRpc in isolate.extensionRPCs) {
-      streamNotify(
+    for (var extensionRpc in _inspector.isolate.extensionRPCs) {
+      _streamNotify(
           'Isolate',
           Event()
             ..kind = EventKind.kServiceExtensionAdded
@@ -179,29 +166,24 @@ class ChromeProxyService implements VmServiceInterface {
   ///
   /// Clears out [_isolate] and all related cached information.
   void destroyIsolate() {
-    if (_isolate == null) return;
-    streamNotify(
+    var isolate = _inspector?.isolate;
+    if (isolate == null) return;
+    _streamNotify(
         'Isolate',
         Event()
           ..kind = EventKind.kIsolateExit
-          ..isolate = toIsolateRef(_isolate));
-    _vm.isolates.removeWhere((ref) => ref.id == _isolate.id);
+          ..isolate = toIsolateRef(isolate));
+    _vm.isolates.removeWhere((ref) => ref.id == isolate.id);
 
-    _isolate = null;
-    _classes.clear();
-    _scriptRefs.clear();
-    _serverPathToScriptRef.clear();
-    _libraries.clear();
-    _libraryRefs.clear();
+    _inspector = null;
     _consoleSubscription.cancel();
     _consoleSubscription = null;
   }
 
   @override
   Future<Breakpoint> addBreakpoint(String isolateId, String scriptId, int line,
-      {int column}) async {
-    return debugger.addBreakpoint(isolateId, scriptId, line, column: column);
-  }
+          {int column}) =>
+      _debugger.addBreakpoint(isolateId, scriptId, line, column: column);
 
   @override
   Future<Breakpoint> addBreakpointAtEntry(String isolateId, String functionId) {
@@ -264,82 +246,9 @@ require("dart_sdk").developer.invokeExtension(
 
   @override
   Future evaluate(String isolateId, String targetId, String expression,
-      {Map<String, String> scope, bool disableBreakpoints}) async {
-    scope ??= {};
-    disableBreakpoints ??= false;
-    var library = await _getLibrary(isolateId, targetId);
-    if (library == null) {
-      throw UnsupportedError(
-          'Evaluate is only supported when `targetId` is a library.');
-    }
-    WipResponse result;
-
-    // If there is scope, we use `callFunctionOn` because that accepts the
-    // `arguments` option.
-    //
-    // If there is no scope we run a normal evaluate call.
-    if (scope.isEmpty) {
-      var evalExpression = '''
-(function() {
-  ${_getLibrarySnippet(library.uri)};
-  return library.$expression;
-})();
-    ''';
-      result = await tabConnection.runtime.sendCommand('Runtime.evaluate',
-          params: {'expression': evalExpression});
-      handleErrorIfPresent(result,
-          evalContents: evalExpression,
-          additionalDetails: {
-            'Dart expression': expression,
-          });
-    } else {
-      var argsString = scope.keys.join(', ');
-      var arguments = scope.values.map((id) => {'objectId': id}).toList();
-      var evalExpression = '''
-function($argsString) {
-  ${_getLibrarySnippet(library.uri)};
-  return library.$expression;
-}
-    ''';
-      result = await tabConnection.runtime
-          .sendCommand('Runtime.callFunctionOn', params: {
-        'functionDeclaration': evalExpression,
-        'arguments': arguments,
-        // TODO(jakemac): Use the executionContext instead, or possibly the
-        // library object. This will get weird if people try to use `this` in
-        // their expression.
-        'objectId': scope.values.first,
-      });
-      handleErrorIfPresent(result,
-          evalContents: evalExpression,
-          additionalDetails: {
-            'Dart expression': expression,
-            'scope': scope,
-          });
-    }
-    var remoteObject =
-        RemoteObject(result.result['result'] as Map<String, dynamic>);
-
-    switch (remoteObject.type) {
-      case 'string':
-        return _primitiveInstance(InstanceKind.kString, remoteObject);
-      case 'number':
-        return _primitiveInstance(InstanceKind.kDouble, remoteObject);
-      case 'boolean':
-        return _primitiveInstance(InstanceKind.kBool, remoteObject);
-      case 'object':
-        return InstanceRef()
-          ..kind = InstanceKind.kPlainInstance
-          ..id = remoteObject.objectId
-          // TODO(jakemac): Create a real ClassRef, we need a way of looking
-          // up the library for a given instance to create it though.
-          // https://github.com/dart-lang/sdk/issues/36771.
-          ..classRef = ClassRef();
-      default:
-        throw UnsupportedError(
-            'Unsupported response type ${remoteObject.type}');
-    }
-  }
+          {Map<String, String> scope, bool disableBreakpoints}) =>
+      _inspector?.evaluate(isolateId, targetId, expression,
+          scope: scope, disableBreakpoints: disableBreakpoints);
 
   @override
   Future evaluateInFrame(String isolateId, int frameIndex, String expression,
@@ -371,140 +280,14 @@ function($argsString) {
   /// Sync version of [getIsolate] for internal use, also has stronger typing
   /// than the public one which has to be dynamic.
   Isolate _getIsolate(String isolateId) {
-    if (_isolate?.id == isolateId) return _isolate;
+    var isolate = _inspector?.isolate;
+    if (isolate?.id == isolateId) return isolate;
     throw ArgumentError.value(
         isolateId, 'isolateId', 'Unrecognized isolate id');
   }
 
   @override
   Future getIsolate(String isolateId) async => _getIsolate(isolateId);
-
-  Future<Library> _constructLibrary(LibraryRef libraryRef) async {
-    // Fetch information about all the classes in this library.
-    var expression = '''
-    (function() {
-    ${_getLibrarySnippet(libraryRef.uri)}
-      var classes = Object.values(library)
-        .filter((l) => l && sdkUtils.isType(l));
-      return classes.map(function(clazz) {
-        var descriptor = {'name': clazz.name};
-
-        // TODO(jakemac): static methods once ddc supports them
-        var methods = sdkUtils.getMethods(clazz);
-        var methodNames = methods ? Object.keys(methods) : [];
-        descriptor['methods'] = {};
-        for (var name of methodNames) {
-          var method = methods[name];
-          descriptor['methods'][name] = {
-            // TODO(jakemac): how can we get actual const info?
-            "isConst": false,
-            "isStatic": false,
-          }
-        }
-
-        // TODO(jakemac): static fields once ddc supports them
-        var fields = sdkUtils.getFields(clazz);
-        var fieldNames = fields ? Object.keys(fields) : [];
-        descriptor['fields'] = {};
-        for (var name of fieldNames) {
-          var field = fields[name];
-          descriptor['fields'][name] = {
-            // TODO(jakemac): how can we get actual const info?
-            "isConst": false,
-            "isFinal": field.isFinal,
-            "isStatic": false,
-          }
-        }
-
-        return descriptor;
-      });
-    })()
-    ''';
-    var classesResult = await tabConnection.runtime.sendCommand(
-        'Runtime.evaluate',
-        params: {'expression': expression, 'returnByValue': true});
-    handleErrorIfPresent(classesResult, evalContents: expression);
-    var classDescriptors = (classesResult.result['result']['value'] as List)
-        .cast<Map<String, Object>>();
-    var classRefs = <ClassRef>[];
-    for (var classDescriptor in classDescriptors) {
-      var name = classDescriptor['name'] as String;
-      var classId = '${libraryRef.id}:$name';
-      var classRef = ClassRef()
-        ..name = name
-        ..id = classId;
-      classRefs.add(classRef);
-
-      var methodRefs = <FuncRef>[];
-      var methodDescriptors =
-          classDescriptor['methods'] as Map<String, dynamic>;
-      methodDescriptors.forEach((name, descriptor) {
-        var methodId = '$classId:$name';
-        methodRefs.add(FuncRef()
-          ..id = methodId
-          ..name = name
-          ..owner = classRef
-          ..isConst = descriptor['isConst'] as bool
-          ..isStatic = descriptor['isStatic'] as bool);
-      });
-
-      var fieldRefs = <FieldRef>[];
-      var fieldDescriptors = classDescriptor['fields'] as Map<String, dynamic>;
-      fieldDescriptors.forEach((name, descriptor) {
-        fieldRefs.add(FieldRef()
-          ..name = name
-          ..declaredType = (InstanceRef()..classRef = ClassRef())
-          ..owner = classRef
-          ..isConst = descriptor['isConst'] as bool
-          ..isFinal = descriptor['isFinal'] as bool
-          ..isStatic = descriptor['isStatic'] as bool);
-      });
-
-      // TODO: Implement the rest of these
-      // https://github.com/dart-lang/webdev/issues/176.
-      _classes[classId] = Class()
-        ..classRef = classRef
-        ..fields = fieldRefs
-        ..functions = methodRefs
-        ..id = classId
-        ..library = libraryRef
-        ..name = name
-        ..interfaces = []
-        ..subclasses = [];
-    }
-
-    // TODO(grouma) - This currently does not support part files.
-    // Figure out how to support them.
-    var scriptRef = ScriptRef()
-      ..uri = libraryRef.uri
-      ..id = createId();
-
-    _scriptRefs[scriptRef.id] = scriptRef;
-    _serverPathToScriptRef[DartUri(libraryRef.id, scriptRef.uri).serverPath] =
-        scriptRef;
-
-    return Library()
-      ..id = libraryRef.id
-      ..name = libraryRef.name
-      ..uri = libraryRef.uri
-      ..classes = classRefs
-      ..debuggable = true
-      ..dependencies = []
-      ..functions = []
-      ..scripts = [scriptRef]
-      ..variables = [];
-  }
-
-  Future<Library> _getLibrary(String isolateId, String objectId) async {
-    if (isolateId != _isolate?.id) return null;
-    var libraryRef = _libraryRefs[objectId];
-    if (libraryRef == null) return null;
-    var library = _libraries[objectId];
-    if (library != null) return library;
-    library = await _constructLibrary(libraryRef);
-    _libraries[objectId] = library;
-    return library;
-  }
 
   @override
   Future<dynamic> getMemoryUsage(String isolateId) async {
@@ -513,57 +296,12 @@ function($argsString) {
 
   @override
   Future getObject(String isolateId, String objectId,
-      {int offset, int count}) async {
-    var library = await _getLibrary(isolateId, objectId);
-    if (library != null) return library;
-    var clazz = _classes[objectId];
-    if (clazz != null) return clazz;
-    var scriptRef = _scriptRefs[objectId];
-    if (scriptRef != null) return await _getScript(isolateId, scriptRef);
-
-    throw UnsupportedError(
-        'Only libraries and classes are supported for getObject');
-  }
+          {int offset, int count}) =>
+      _inspector?.getObject(isolateId, objectId, offset: offset, count: count);
 
   @override
-  Future<ScriptList> getScripts(String isolateId) async {
-    var scripts = await scriptRefs(isolateId);
-    return ScriptList()..scripts = scripts;
-  }
-
-  /// Returns the [ScriptRef] for the provided Dart server path [uri].
-  ScriptRef scriptRefFor(String uri) => _serverPathToScriptRef[uri];
-
-  Future<Script> _getScript(String isolateId, ScriptRef scriptRef) async {
-    var libraryId = scriptRef.uri;
-    // TODO(401): Remove uri parameter.
-    var serverPath = DartUri(libraryId, uri).serverPath;
-    var script = await assetHandler(serverPath);
-    return Script()
-      ..library = _libraryRefs[libraryId]
-      ..id = scriptRef.id
-      ..uri = libraryId
-      ..tokenPosTable = debugger.sources.tokenPosTableFor(serverPath)
-      ..source = script;
-  }
-
-  /// Internal method to list all scripts in the isolate.
-  ///
-  /// This is used as the underlying mechanism
-  /// for [getScripts].
-  Future<List<ScriptRef>> scriptRefs(String isolateId) async {
-    var isolate = _getIsolate(isolateId);
-    var scripts = <ScriptRef>[];
-    for (var lib in isolate.libraries) {
-      // We can't provide the source for `dart:` imports so ignore for now.
-      // Also `main.dart.bootstrap` does not have a corresponding script.
-      if (lib.id.startsWith('dart:') || lib.id.endsWith('.bootstrap')) continue;
-      scripts.addAll((await _getLibrary(isolateId, lib.id)).scripts);
-    }
-    // TODO(alanknight): Is this the same as the values in _scriptRefs after
-    // constructing all the libraries? Make that clearer.
-    return scripts;
-  }
+  Future<ScriptList> getScripts(String isolateId) =>
+      _inspector?.getScripts(isolateId);
 
   @override
   Future<SourceReport> getSourceReport(String isolateId, List<String> reports,
@@ -575,7 +313,7 @@ function($argsString) {
   ///
   /// Returns null if the corresponding isolate is not paused.
   @override
-  Future<Stack> getStack(String isolateId) => debugger.getStack(isolateId);
+  Future<Stack> getStack(String isolateId) => _debugger.getStack(isolateId);
 
   @override
   Future<VM> getVM() => Future.value(_vm);
@@ -636,7 +374,7 @@ function($argsString) {
   }
 
   @override
-  Future<Success> pause(String isolateId) => debugger.pause();
+  Future<Success> pause(String isolateId) => _debugger.pause();
 
   @override
   Future<Success> registerService(String service, String alias) async {
@@ -650,11 +388,8 @@ function($argsString) {
   }
 
   @override
-  Future<Success> removeBreakpoint(
-      String isolateId, String breakpointId) async {
-    await debugger.removeBreakpoint(isolateId, breakpointId);
-    return Success();
-  }
+  Future<Success> removeBreakpoint(String isolateId, String breakpointId) =>
+      _debugger.removeBreakpoint(isolateId, breakpointId);
 
   @override
   Future<Success> requestHeapSnapshot(
@@ -665,20 +400,11 @@ function($argsString) {
   @override
   Future<Success> resume(String isolateId,
           {String step, int frameIndex}) async =>
-      debugger.resume(isolateId, step: step, frameIndex: frameIndex);
+      _debugger.resume(isolateId, step: step, frameIndex: frameIndex);
 
   @override
-  Future<Success> setExceptionPauseMode(String isolateId, String mode) async {
-    // Validates that the ID is correct.
-    _getIsolate(isolateId);
-
-    var pauseState = _pauseModePauseStates[mode.toLowerCase()] ??
-        // ignore: only_throw_errors
-        (throw RPCError(
-            'setExceptionPauseMode', -32602, 'Unsupported mode `$mode`'));
-    await tabConnection.debugger.setPauseOnExceptions(pauseState);
-    return Success();
-  }
+  Future<Success> setExceptionPauseMode(String isolateId, String mode) =>
+      _debugger.setExceptionPauseMode(isolateId, mode);
 
   @override
   Future<Success> setFlag(String name, String value) {
@@ -701,7 +427,7 @@ function($argsString) {
   @override
   Future<Success> setVMName(String name) async {
     _vm.name = name;
-    streamNotify(
+    _streamNotify(
         'VM',
         Event()
           ..kind = EventKind.kVMUpdate
@@ -742,24 +468,26 @@ function($argsString) {
     }, onListen: () {
       chromeConsoleSubscription =
           tabConnection.runtime.onConsoleAPICalled.listen((e) {
-        if (_isolate == null) return;
+        var isolate = _inspector?.isolate;
+        if (isolate == null) return;
         if (!filter(e)) return;
         var args = e.params['args'] as List;
         var item = args[0] as Map;
         var value = '${item["value"]}\n';
         controller.add(Event()
           ..kind = EventKind.kWriteEvent
-          ..isolate = toIsolateRef(_isolate)
+          ..isolate = toIsolateRef(isolate)
           ..bytes = base64.encode(utf8.encode(value))
           ..timestamp = e.timestamp.toInt());
       });
       if (includeExceptions) {
         exceptionsSubscription =
             tabConnection.runtime.onExceptionThrown.listen((e) {
-          if (_isolate == null) return;
+          var isolate = _inspector?.isolate;
+          if (isolate == null) return;
           controller.add(Event()
             ..kind = EventKind.kWriteEvent
-            ..isolate = toIsolateRef(_isolate)
+            ..isolate = toIsolateRef(isolate)
             ..bytes = base64
                 .encode(utf8.encode(e.exceptionDetails.exception.description)));
         });
@@ -768,38 +496,19 @@ function($argsString) {
     return controller;
   }
 
-  /// Runs an eval on the page to compute all existing registered extensions.
-  Future<List<String>> _getExtensionRpcs() async {
-    var expression = "require('dart_sdk').developer._extensions.keys.toList();";
-    var extensionsResult = await tabConnection.runtime.sendCommand(
-        'Runtime.evaluate',
-        params: {'expression': expression, 'returnByValue': true});
-    handleErrorIfPresent(extensionsResult, evalContents: expression);
-    return List.from(extensionsResult.result['result']['value'] as List);
-  }
-
-  /// Runs an eval on the page to compute all the library names in the app.
-  Future<List<String>> _getLibraryNames() async {
-    var expression = "require('dart_sdk').dart.getLibraries();";
-    var librariesResult = await tabConnection.runtime.sendCommand(
-        'Runtime.evaluate',
-        params: {'expression': expression, 'returnByValue': true});
-    handleErrorIfPresent(librariesResult, evalContents: expression);
-    return List.from(librariesResult.result['result']['value'] as List);
-  }
-
   /// Listens for chrome console events and handles the ones we care about.
   void _setUpChromeConsoleListeners(IsolateRef isolateRef) {
     _consoleSubscription = tabConnection.runtime.onConsoleAPICalled
         .listen((ConsoleAPIEvent event) {
-      if (_isolate == null) return;
+      var isolate = _inspector?.isolate;
+      if (isolate == null) return;
       if (event.type != 'debug') return;
       var firstArgValue = event.args[0].value as String;
       switch (firstArgValue) {
         case 'dart.developer.registerExtension':
           var service = event.args[1].value as String;
-          _isolate.extensionRPCs.add(service);
-          streamNotify(
+          isolate.extensionRPCs.add(service);
+          _streamNotify(
               'Isolate',
               Event()
                 ..kind = EventKind.kServiceExtensionAdded
@@ -807,7 +516,7 @@ function($argsString) {
                 ..isolate = isolateRef);
           break;
         case 'dart.developer.postEvent':
-          streamNotify(
+          _streamNotify(
               'Extension',
               Event()
                 ..kind = EventKind.kExtension
@@ -827,7 +536,7 @@ function($argsString) {
             // up the library for a given instance to create it though.
             // https://github.com/dart-lang/sdk/issues/36771.
             ..classRef = ClassRef();
-          streamNotify(
+          _streamNotify(
               'Debug',
               Event()
                 ..kind = EventKind.kInspect
@@ -841,9 +550,7 @@ function($argsString) {
     });
   }
 
-  /// Adds [event] to the stream with [streamId] if there is anybody listening
-  /// on that stream.
-  void streamNotify(String streamId, Event event) {
+  void _streamNotify(String streamId, Event event) {
     var controller = _streamControllers[streamId];
     if (controller == null) return;
     controller.add(event);
@@ -867,25 +574,6 @@ void handleErrorIfPresent(WipResponse response,
         additionalDetails: additionalDetails);
   }
 }
-
-/// Creates a snippet of JS code that initializes a `library` variable that has
-/// the actual library object in DDC for [libraryUri].
-///
-/// In DDC we have module libraries indexed by names of the form
-/// 'packages/package/mainFile' with no .dart suffix on the file, or
-/// 'directory/packageName/mainFile', also with no .dart suffix, and relative to
-/// the serving root, normally /web within the package. These modules have a map
-/// from the URI with a Dart-specific scheme (package: or org-dartlang-app:) to
-/// the library objects. The [libraryUri] parameter should be one of these
-/// Dart-specific scheme URIs, and we set `library` the corresponding library.
-String _getLibrarySnippet(String libraryUri) => '''
-  var libraryName = '$libraryUri';
-  var sdkUtils = require('dart_sdk').dart;
-  var moduleName = sdkUtils.getModuleNames().find(
-    (name) => sdkUtils.getModuleLibraries(name)[libraryName]);
-  var library = sdkUtils.getModuleLibraries(moduleName)[libraryName];
-  if (!library) throw 'cannot find library for ' + libraryName + ' under ' + moduleName;
-''';
 
 class ChromeDebugException extends ExceptionDetails implements Exception {
   /// Optional, additional information about the exception.
@@ -919,22 +607,4 @@ class ChromeDebugException extends ExceptionDetails implements Exception {
     }
     return description.toString();
   }
-}
-
-/// Converts from ExceptionPauseMode strings to [PauseState] enums.
-const _pauseModePauseStates = {
-  'none': PauseState.none,
-  'all': PauseState.all,
-  'unhandled': PauseState.uncaught,
-};
-
-/// Creates an [InstanceRef] for a primitive [RemoteObject].
-InstanceRef _primitiveInstance(String kind, RemoteObject remoteObject) {
-  var classRef = ClassRef()
-    ..id = 'dart:core:${remoteObject.type}'
-    ..name = kind;
-  return InstanceRef()
-    ..valueAsString = '${remoteObject.value}'
-    ..classRef = classRef
-    ..kind = kind;
 }
